feat

+++ AutoEncoderTrainer.py
@@ -70,12 +70,12 @@ if __name__ == "__main__":
 
     if "demo" == cfg.strategy:
         trainer = pl.Trainer( accelerator="gpu", devices=1, logger=logger, callbacks=[cb,mc],
-                              accumulate_grad_batches=cfg.accumulate_grad_batches)
+                              accumulate_grad_batches=cfg.accumulate_grad_batches, precision=16)
     else:
         #in case you wanna train it on multi gpu
         trainer = pl.Trainer( accelerator="gpu", devices=2, logger=logger,callbacks=[cb,mc],
                              accumulate_grad_batches=cfg.accumulate_grad_batches,
-                              strategy=DDPStrategy(find_unused_parameters=True))
+                              strategy=DDPStrategy(find_unused_parameters=True), precision=16)
     trainer.fit(model, train_dataloaders=gdm.train_dataloader(), val_dataloaders=gdm.val_dataloader(),
                 ckpt_path=checkpoint_path)
